release 3 method getAvailableRooms implemented

+++ backend-demo/src/main/java/de/ait/todo/services/impl/RoomsServiceImpl.java
@@ -3,14 +3,19 @@ package de.ait.todo.services.impl;
 import de.ait.todo.dto.NewRoomDto;
 import de.ait.todo.dto.RoomDto;
 import de.ait.todo.dto.RoomsPage;
+import de.ait.todo.exceptions.IncorrectDeleteException;
 import de.ait.todo.exceptions.NotFoundException;
+import de.ait.todo.models.Booking;
 import de.ait.todo.models.Room;
+import de.ait.todo.repositories.BookingsRepository;
 import de.ait.todo.repositories.RoomsRepository;
 import de.ait.todo.services.RoomsService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 
+import java.time.LocalDate;
+import java.util.ArrayList;
 import java.util.List;
 
 import static de.ait.todo.dto.UserDto.from;
@@ -22,6 +27,8 @@ public class RoomsServiceImpl implements RoomsService {
 
 
     private final RoomsRepository roomsRepository;
+
+    private final BookingsRepository bookingsRepository;
 
     @Override
     public int createRoom(NewRoomDto newRoom) {
@@ -80,10 +87,48 @@ public class RoomsServiceImpl implements RoomsService {
                        .orElseThrow( ()->
                                new NotFoundException("room with id <" + id + "> not found")
                        );
+       if(room.isBooked()){
+           throw new IncorrectDeleteException("Cannot delete a booked room");
+       }
 
         roomsRepository.deleteById(id);
 
         return RoomDto.from(room);
     }
 
+    @Override
+    public List<RoomDto> getAvailableRooms(LocalDate startDate, LocalDate endDate) {
+
+        List<Room> allRooms = roomsRepository.findAll();
+
+
+        List<Room> availableRooms = new ArrayList<>();
+
+
+
+        for (Room room : allRooms) {
+            boolean isAvailable = true;
+            List<Booking> allBookings = bookingsRepository.findAllByRoomsContains(room);
+
+            // Проверка наличия брони для комнаты в заданный период
+            for (Booking currentBooking : allBookings) {
+                if (((currentBooking.getCheCkIn().isEqual(startDate) || currentBooking.getCheCkIn().isAfter(startDate)) && (currentBooking.getCheCkIn().isEqual(endDate) || currentBooking.getCheCkIn().isBefore(endDate))) ||
+                        ((currentBooking.getCheckOut().isEqual(startDate) || currentBooking.getCheckOut().isAfter(startDate)) && (currentBooking.getCheckOut().isEqual(endDate) || currentBooking.getCheckOut().isBefore(endDate)))) {
+                    // Промежуток дат пересекается с startBookingDate
+
+                    isAvailable = false;
+                    System.out.println("Промежуток дат пересекается");
+                    break;
+                } else {
+                    // Промежуток дат не пересекается с startBookingDate
+                    System.out.println("Промежуток дат не пересекается");
+                }
+            }
+
+            if (isAvailable) {
+                availableRooms.add(room);
+            }
+        }
+        return RoomDto.from(availableRooms);
+    }
 }
